+++ mail/proton-mail-common/src/user_context.rs
@@ -45,7 +45,7 @@ impl MailUserContext {
     ) -> MailContextResult<Arc<Self>> {
         let stash = user_context.stash().clone();
         let cache_path = mail_context.mail_cache_path(user_context.user_id());
-        let cache = Cache::new(cache_path, mail_context.mail_cache_size)?;
+        let cache = Cache::new(cache_path, mail_context.mail_cache_size, &stash).await?;
         let action_queue = new_action_queue(stash).await.unwrap();
         let this = Arc::new_cyclic(|this| Self {
             this: Weak::clone(this),